Modification in data type of Paramters in methods

+++ src/main/java/com/ecommerce/dao/DAOAddress.java
@@ -30,7 +30,7 @@ public class DAOAddress {
         }
     }
 
-    public Address getAddressById(Long id) {
+    public Address getAddressById(int id) {
         try (Session session = sessionFactory.openSession()) {
             return session.get(Address.class, id);
         }
